+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 更改日志 | Change Log
 
 - 自定义常量联合类型定义名 | Custom constant union type definition name
+- 添加 API `/api/system/getConf` | Add API `/api/system/getConf`
 
 ## v0.3.2 / 2023-09-19